+++ cfdtools/data.py
@@ -3,12 +3,12 @@ import numpy.fft as fftm
 import matplotlib.pyplot as plt
 import cfdtools.api as api
 import cfdtools.hdf5 as hdf5
+import cfdtools.meshbase as meshbase
 
 
 class DataSetBase:
-
     _available_Xrep = ('nodal', 'cellaverage')
-    _available_Trep = 'instant'
+    _available_Trep = ('instant',)
 
     def __init__(self, Xrep='cellaverage', ndof=1, Trep='instant'):
         self.Trep = Trep
@@ -59,13 +59,18 @@ class DataSetBase:
         return self._geoprop[key] if key else self._geoprop
 
     def set_nodes(self, x, y=None, z=None):
+        """ """
         self._geoprop['x'] = x
-        if y: self._geoprop['y'] = y
-        if z: self._geoprop['z'] = z
+        if y:
+            self._geoprop['y'] = y
+        if z:
+            self._geoprop['z'] = z
+
+    def set_mesh(self, mesh):
+        self._geoprop['mesh'] = mesh
 
 
 class DataSet(DataSetBase):
-
     _available_Xrep = ('nodal', 'cellaverage', 'spectralcell')
     _available_Trep = ('instant', 'timeevol', 'spectrum', 'spectrogram')
 
@@ -105,17 +110,15 @@ class DataSet(DataSetBase):
 
     def dataSet_spectrum(self, datafilter=None, imin=None, imax=None):
         dtavg, dtdev, n = self.dtstats(imin, imax)  # check 'timeevol'
-        if dtdev / dtavg > 1.e-6:
-            api.io.warning(
-                f"dt standard deviation is significant: {dtdev / dtavg}"
-            )
+        if dtdev / dtavg > 1.0e-6:
+            api.io.warning(f"dt standard deviation is significant: {dtdev / dtavg}")
         f = fftm.fftfreq(n, dtavg)
         newdataset = DataSet(self.Xrep, self.ndof, Trep='spectrum')
         newdataset.add_data('freq', f)
         datalist = datafilter if datafilter else list(self.keys())
         datalist.remove('time')
         for name in datalist:
-            newdataset.add_data(name, np.abs(fftm.fft(self[name]-dtavg)))
+            newdataset.add_data(name, np.abs(fftm.fft(self[name])))
         return newdataset
 
     def dataSet_spectrogram(self, datafilter=None, window=None):
@@ -123,26 +126,26 @@ class DataSet(DataSetBase):
         datalist = datafilter if datafilter else list(self.keys())
         datalist.remove('time')
         dtavg, dtdev, ntot = self.dtstats()  # check 'timeevol'
-        # print(dtavg, dtdev)
-        if dtdev / dtavg > 1.e-6:
-            api.io.warning(
-                f"dt standard deviation is significant: {100*dtdev / dtavg:.4f}%"
-            )
+        if dtdev / dtavg > 1.0e-6:
+            api.io.warning(f"dt standard deviation is significant: {100*dtdev / dtavg:.4f}%")
         newdataset = DataSet(self.Xrep, self.ndof, Trep='spectrogram')
-        nwind = window if window else int(ntot/100)
-        hwin = int(nwind/2)
+        nwind = window if window else int(ntot / 100)
+        hwin = int(nwind / 2)
         f = fftm.fftfreq(nwind, dtavg)
-        newdataset.add_data('freq', f[lowcrop:hwin+1])
-        nit = int(2*ntot/nwind)-1
+        newdataset.add_data('freq', f[lowcrop : hwin + 1])
+        nit = int(2 * ntot / nwind) - 1
         newdataset.add_data('time', (self['time'][hwin::hwin])[:-1])
         for name in datalist:
-            v = np.zeros((nit, hwin+1-lowcrop))
+            v = np.zeros((nit, hwin + 1 - lowcrop))
             # print(ntot, nit, nwind, hwin, newdataset['freq'].shape, f.shape)
             for i in range(nit):
-                v[i,:] = np.log(np.abs(fftm.rfft(self[name][hwin*i:hwin*(i+2)+1]*np.hanning(2*hwin+1))))[lowcrop:]  
+                v[i, :] = np.log(
+                    np.abs(fftm.rfft(self[name][hwin * i : hwin * (i + 2) + 1] * np.hanning(2 * hwin + 1)))
+                )[lowcrop:]
             # print(v.shape, newdataset['freq'].shape, newdataset['time'].shape)
             newdataset.add_data(name, v)
         return newdataset
+
 
 # class manufactured_DataSet(DataSet):
 #
@@ -151,9 +154,11 @@ class DataSet(DataSetBase):
 
 
 class DataSetList(DataSetBase):
-
+    _version = 1
     _available_Xrep = ('nodal', 'cellaverage', 'spectralcell')
     _available_Trep = ('instant', 'timeevol', 'pod')
+    # defines the names of data items that should be written as attributes; not hdf5 dataset
+    _property_names = ('time',)
 
     def __init__(self, ndataset, Xrep='cellaverage', ndof=1, Trep='timeevol'):
         super().__init__(Xrep, ndof, Trep)
@@ -169,8 +174,7 @@ class DataSetList(DataSetBase):
         datalist = datafilter if datafilter else list(self.keys())
         datalist.remove('time')
         dtavg, dtdev, ntot = self.dtstats()  # check 'timeevol'
-        # print(dtavg, dtdev)
-        if dtdev / dtavg > 1.e-6:
+        if dtdev / dtavg > 1.0e-6:
             api.io.warning(f"dt standard deviation is significant: {100*dtdev / dtavg:.4f}%")
         newdataset = DataSet(self.Xrep, self.ndof, Trep='spectrogram')
         return newdataset
@@ -181,8 +185,64 @@ class DataSetList(DataSetBase):
         Args:
             hgroup (hdf5.Group): _description_
         """
-        # ? n = len(self._datalist)
         for i, datadict in enumerate(self._datalist):
             datagroup = hgroup.create_group(f"i{i:06}")
             for vname, var in datadict.items():
-                datagroup.create_dataset(vname, data=var, **options)
+                if vname in self._property_names:
+                    datagroup.attrs[vname] = var
+                else:
+                    datagroup.create_dataset(vname, data=var, **options)
+
+    def xdmf_content(self, filename, geometry_content):
+        """Create the XDMF content associated to the data set.
+
+        :param str filename: Name of the output HDF5 file.
+        :param list(str) geometry_content: XDMF content for the geometry and topology of the mesh.
+          The mesh is time-invariant.
+        :return: The XDMF content.
+        :rtype: list(str)
+        """
+        lines = ['<Grid Name="IC3" GridType="Collection" CollectionType="Temporal">']
+
+        for i, datadict in enumerate(self._datalist):
+            lines += ['<Grid Name="Unstructured Mesh">']
+            lines += [f'<Time Value="{i}"/>']
+            lines += geometry_content
+            data_wo_properties = {
+                vname: datadict[vname] for vname in datadict if vname not in self._property_names
+            }
+            for vname, var in data_wo_properties.items():
+                if len(var.shape) == 1:
+                    typ = "Scalar"
+                elif len(var.shape) == 2:
+                    typ = "Vector"
+                else:
+                    raise ValueError(f"Type of data cannot be deduced with shape dimension {len(var.shape)}")
+                dim = " ".join(str(j) for j in var.shape)
+
+                lines += [f'<Attribute AttributeType="{typ}" Center="Cell" Name="{vname}">']
+                lines += [f'<DataItem Dimensions="{dim}" Format="HDF">']
+                lines += [f"{filename}:/datalist/i{i:06}/{vname}"]
+                lines += ["</DataItem>"]
+                lines += ["</Attribute>"]
+            lines += ["</Grid>"]
+        lines += ["</Grid>"]
+
+        return lines
+
+    def dumphdf(self, filename, overwrite=False, **options):
+        file = hdf5.h5File(filename)
+        if not overwrite:
+            file.find_safe_newfile()
+        file.open(mode="w", datatype='datalist', version=self._version)
+        # map element types from vtk type to cfdtools type
+        if isinstance(self._geoprop['mesh'], meshbase.Mesh):
+            celldict = dict(self._geoprop['mesh']._cell2node)
+        else:
+            api.error_stop("not yet implemented to dump mesh with instance {self._geoprop['mesh']}")
+        file.write_unsmesh(celldict, self._geoprop['mesh'].nodescoord(ndarray=True), **options)
+        #
+        hdata = file._h5file.create_group("datalist")
+        self._dumphdfgroup(hdata, **options)
+        file.close()
+        return file.filename