+++ cfdtools/meshbase/__init__.py
@@ -1,1 +1,1 @@
-from cfdtools.meshbase._mesh import Mesh # make it available from cfdtools.meshbase
+from cfdtools.meshbase._mesh import Mesh  # make it available from cfdtools.meshbase