+++ cfdtools/vtk/_vtk.py
@@ -1,3 +1,5 @@
+import os
+
 try:
     import pyvista as pv
     from pyvista import CellType
@@ -6,15 +8,10 @@ try:
 except ImportError:
     importpyvista = False
 import cfdtools.api as api
-import cfdtools.utils.maths as maths
-import cfdtools.meshbase._elements as ele
+import cfdtools.meshbase._connectivity as _conn
 import cfdtools.meshbase._mesh as cfdmesh
-from cfdtools.data import DataSetList
 import cfdtools.hdf5 as hdf5
-from pathlib import Path
 import numpy as np
-import scipy.spatial as spspa
-
 
 try:
     vtktype_ele = {
@@ -23,16 +20,28 @@ try:
         'hexa8': CellType.HEXAHEDRON,
     }
     ele_vtktype = {i: etype for etype, i in vtktype_ele.items()}
+    PYVISTA2XDMF = {
+        CellType.HEXAHEDRON: "Hexahedron",
+        CellType.QUAD: "Quadrilateral",
+    }
 except NameError:
-    api.io.print('error', "pyvista (with CellType) could not be imported")
+    api.error_stop("pyvista (with CellType) could not be imported")
     raise
 
 
 @api.fileformat_writer("VTK", '.vtu')
 class vtkMesh:
-    '''Implementation of the writer to write binary Vtk files using pyvista'''
+    """Implementation of the writer to write binary Vtk files using pyvista"""
+
+    _version = 1
 
     def __init__(self, mesh=None, pvmesh=None):
+        """Initialisation of vtkMesh, encapsulation of pyvista vtk mesh
+
+        Args:
+            mesh (meshbase, optional): cfdtools meshbase object. Defaults to None.
+            pvmesh (pyvista, optional): direct pyvista mesh. Defaults to None.
+        """
         if mesh and pvmesh:
             api.error_stop("vtkMesh cannot be initialized by both structures")
         if mesh:
@@ -47,14 +56,10 @@ class vtkMesh:
         self._reset()
         self._mesh = mesh
         coords = self._mesh.nodescoord(ndarray=True)
-        try:
-            self._celldict = {
-                vtktype_ele[etype]: elem2node['elem2node']
-                for etype, elem2node in self._mesh._cell2node.items()
-            }
-            self._grid = pv.UnstructuredGrid(self._celldict, coords)
-        except NameError:
-            api.error_stop("pyvista (with CellType) could not be imported")
+        self._celldict = {
+            vtktype_ele[etype]: elem2node['elem2node'] for etype, elem2node in self._mesh._cell2node.items()
+        }
+        self._grid = pv.UnstructuredGrid(self._celldict, coords)
 
     def set_pvmesh(self, pvmesh: cfdmesh.Mesh):
         self._reset()
@@ -65,11 +70,33 @@ class vtkMesh:
 
     def read(self, filename):
         self.__init__()
-        self._grid = pv.read(filename)
+        self.set_pvmesh(pv.read(filename))
+
+    def export_mesh(self):
+        """generates a cfdtools meshbase from vtk connectivity, boundary conditions are missing
+
+        Returns:
+            meshbase.Mesh: cell to node connectivity and node positions
+        """
+        cellnode = _conn.elem_connectivity()
+        for itype, cellco in self._grid.cells_dict.items():
+            cellnode.add_elems(ele_vtktype[itype], cellco)
+        mesh = cfdmesh.Mesh(ncell=self.ncell, nnode=self.nnode)
+        mesh.set_cell2node(cellnode)
+        mesh.set_nodescoord_nd(self._grid.points)
+        return mesh
 
     @property
     def pyvista_grid(self):
         return self._grid
+
+    @property
+    def ncell(self):
+        return self._grid.n_cells
+
+    @property
+    def nnode(self):
+        return self._grid.n_points
 
     def brief(self):
         if self._grid:
@@ -91,118 +118,60 @@ class vtkMesh:
         self.pyvista_grid().plot(background=background, show_edges=show_edges, *args, **kwargs)
 
     def importhdfgroup(self, hgroup: hdf5.Group, verbose=False):
-        assert hgroup.attrs['meshtype'] == 'unsvtk'
+        assert hgroup.attrs['meshtype'] in ('unsmesh',)
         points = np.array(hgroup["nodes"])
         celldict = {vtktype_ele[etype]: np.array(elem2node) for etype, elem2node in hgroup["cells"].items()}
         self.set_pvmesh(pv.UnstructuredGrid(celldict, points))
 
-    def dumhdfgroup(self, hgroup: hdf5.Group, **options):
-        hgroup.attrs['meshtype'] = 'unsvtk'
-        hgroup.create_dataset("nodes", data=self._grid.points, **options)
-        hcells = hgroup.create_group("cells")
+    def xdmf_content(self, filename):
+        """Create the XDMF content associated to the mesh.
+
+        :param str filename: Name of the output HDF5 file.
+        :return: The XDMF content.
+        :rtype: list(str)
+        """
+        dim2type = {1: "X", 2: "XY", 3: "XYZ"}
+        geomtype = dim2type[self._grid.points.shape[1]]
+        lines = [f'<Geometry GeometryType="{geomtype}">']
+        lines += [f'<DataItem Dimensions="{np.prod(self._grid.points.shape)}" Format="HDF">']
+        lines += [f"{filename}:/mesh/nodes"]
+        lines += ["</DataItem>"]
+        lines += ["</Geometry>"]
+
         for itype, cellco in self._grid.cells_dict.items():
-            hcells.create_dataset(ele_vtktype[itype], data=cellco, **options)
+            lines += [
+                f'<Topology NumberOfElements="{cellco.shape[0]:d}" TopologyType="{PYVISTA2XDMF[itype]}">'
+            ]
+            lines += [f'<DataItem Dimensions="{np.prod(cellco.shape)}" Format="HDF">']
+            lines += [f"{filename}:/mesh/cells/{ele_vtktype[itype]}"]
+            lines += ["</DataItem>"]
+            lines += ["</Topology>"]
+
+        return lines
 
     def volumes(self):
         if not self._volume:
             self._volume = self._grid.compute_cell_sizes().cell_data["Volume"]
         return self._volume
 
+    def dumphdf(self, filename, overwrite=False, **options):
+        """dump pyvista mesh and (future) data to cfdtools hdf5 file
+        (in the future) must be converted to dataset and dump done by dataset
 
-class vtkList:
-    def __init__(self, filelist, verbose=False) -> None:
-        self._list = filelist
-        self._verbose = verbose
+        Args:
+            filename (string or pathlib): file name
+            overwrite (bool): overwrite file or find new safe name
+            **options: passed to dumpdfgroup and h5py.createdataset
 
-    @property
-    def nfile(self):
-        return len(self._list)
-
-    def allexist(self):
-        return all(Path(file).exists() for file in self._list)
-
-    def check_order(self, pos='cellcenter', tol=1e-10):
-        mappos = {
-            'cellcenter': lambda m: m.cell_centers().points,
-            'node': lambda m: m.points,
-        }
-        count = 0
-        assert self.nfile > 0
-        ref = mappos[pos](pv.read(self._list[0]))
-        for name in self._list:
-            mesh = pv.read(name)
-            d = np.max(maths.distance(ref, mappos[pos](mesh)))
-            if d > tol:
-                count += 1
-                if self._verbose:
-                    api.io.printstd(f"  . {name}: {d}")
-        if self._verbose:
-            api.io.printstd(f"  {count}/{self.nfile} grids are not {pos}-coincident")
-        return count == 0
-
-    def read(self, filterdata=None, reorder=False, tol=1e-10):
-        count = 0
-        assert self.nfile > 0
-        Tread = api.Timer()
-        Tcomp = api.Timer()
-        Tsort = api.Timer()
-        Tread.start()
-        self._mesh = pv.read(self._list[0])
-        self._ncell = self._mesh.n_cells
-        ctrRef = self._mesh.cell_centers().points
-        Tread.pause()
-        if self._verbose:
-            api.io.printstd("> build kd-tree")
-        Tsort.start()
-        tree = spspa.KDTree(ctrRef)
-        Tsort.pause()
-        #
-        self._data = DataSetList(self.nfile, Xrep='cellaverage', Trep='instant')
-        # may add alive-progress or other
-        if self._verbose:
-            api.io.printstd("> read all files, get only CELL data")
-        if filterdata:
-            api.io.printstd(f"  select only {', '.join(filterdata)}")
-        for name in self._list:
-            if self._verbose:
-                api.io.printstd(f"  - {name}")
-            Tread.start()
-            vtk = pv.read(name)
-            Tread.pause()
-            namelist = filterdata if filterdata else vtk.cell_data.keys()
-            Tcomp.start()
-            ctr = vtk.cell_centers().points
-            d = np.max(maths.distance(ctrRef, ctr))
-            Tcomp.pause()
-            if d > tol:
-                count += 1
-                # if self._verbose:
-                #     api.io.printstd(f"  . {name}: {d}")
-                Tsort.start()
-                dfinal, index = tree.query(ctr, p=2)
-                # reverse indexing to sort new arrays
-                rindex = index.copy()
-                rindex[index] = np.arange(index.size)
-                assert np.max(dfinal) <= tol
-                # automatically deals with differnt shapes
-                datalist = {name: vtk.cell_data[name][rindex] for name in namelist}
-                Tsort.pause()
-                self._data.add_datalist(datalist)
-        if self._verbose:
-            api.io.printstd(f"  {count}/{self.nfile} grids were not coincident")
-            api.io.printstd(f"       file reading: {Tread.elapsed:.2f}s")
-            api.io.printstd(f"    grid comparison: {Tcomp.elapsed:.2f}s")
-            api.io.printstd(f"    data reordering: {Tsort.elapsed:.2f}s")
-
-    def dumphdf(self, filename, **options):
+        Returns:
+            string: file name of the actual (safe) saved file
+        """
         file = hdf5.h5File(filename)
-        file.find_safe_newfile()
-        file.open(mode="w", datatype='datalist')
-        hmesh = file._h5file.create_group("mesh")
-        vtkmesh = vtkMesh(pvmesh=self._mesh)
-        vtkmesh.dumhdfgroup(hmesh, **options)
-        #
-        hdata = file._h5file.create_group("datalist")
-        self._data._dumphdfgroup(hdata, **options)
+        if not overwrite:
+            file.find_safe_newfile()
+        file.open(mode="w", datatype='dataset', version=self._version)
+        # map element types from vtk type to cfdtools type
+        celldict = {ele_vtktype[itype]: cellco for itype, cellco in self._grid.cells_dict.items()}
+        file.write_unsmesh(celldict, self._grid.points, **options)
         file.close()
         return file.filename